refactor: extract broker, compass, webhook, health http features into per-feature crates
