--- api/src/application/http/webhook/errors.rs
@@ -1,1 +1,0 @@
-
