+++ libs/ferriskey-api-compass/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "ferriskey-api-compass"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+
+[dependencies]
+ferriskey-api-core = { path = "../ferriskey-api-core" }
+ferriskey-core = { path = "../../core" }
+axum = "0.8.8"
+serde = "1.0.228"
+serde_json = "1.0.148"
+utoipa = { version = "5.4.0", features = ["chrono", "uuid"] }
+uuid = { version = "1.19.0", features = ["serde", "v4", "v7"] }
+chrono = { version = "0.4.40", features = ["serde"] }
