+++ libs/ferriskey-api-health/src/handlers.rs
