+++ libs/ferriskey-api-webhook/src/handlers/create_webhook.rs
@@ -1,0 +1,61 @@
+use crate::validators::CreateWebhookValidator;
+use axum::{
+    Extension,
+    extract::{Path, State},
+};
+use ferriskey_api_core::api_entities::api_error::{ApiError, ApiErrorResponse, ValidateJson};
+use ferriskey_api_core::api_entities::response::Response;
+use ferriskey_api_core::app_state::AppState;
+use ferriskey_core::domain::authentication::value_objects::Identity;
+use ferriskey_core::domain::webhook::entities::webhook::Webhook;
+use ferriskey_core::domain::webhook::ports::{CreateWebhookInput, WebhookService};
+use serde::{Deserialize, Serialize};
+use utoipa::ToSchema;
+
+#[derive(Debug, Serialize, Deserialize, ToSchema, PartialEq)]
+pub struct CreateWebhookResponse {
+    pub data: Webhook,
+}
+
+#[utoipa::path(
+    post,
+    path = "",
+    tag = "webhook",
+    summary = "Create webhook",
+    description = "Creates a new webhook in the system related to the current realm.",
+    params(
+        ("realm_name" = String, Path, description = "Realm name"),
+    ),
+    request_body = CreateWebhookValidator,
+    responses(
+        (status = 200, description = "Webhook created successfully", body = CreateWebhookResponse),
+        (status = 400, description = "Invalid request data", body = ApiErrorResponse),
+        (status = 401, description = "Realm not found", body = ApiErrorResponse),
+        (status = 403, description = "Insufficient permissions", body = ApiErrorResponse),
+        (status = 500, description = "Internal server error", body = ApiErrorResponse),
+    ),
+)]
+pub async fn create_webhook(
+    Path(realm_name): Path<String>,
+    State(state): State<AppState>,
+    Extension(identity): Extension<Identity>,
+    ValidateJson(payload): ValidateJson<CreateWebhookValidator>,
+) -> Result<Response<CreateWebhookResponse>, ApiError> {
+    let webhook = state
+        .service
+        .create_webhook(
+            identity,
+            CreateWebhookInput {
+                realm_name,
+                name: payload.name,
+                description: payload.description,
+                endpoint: payload.endpoint,
+                headers: payload.headers,
+                subscribers: payload.subscribers,
+            },
+        )
+        .await
+        .map_err(ApiError::from)?;
+
+    Ok(Response::OK(CreateWebhookResponse { data: webhook }))
+}
