+++ libs/ferriskey-api-webhook/src/handlers/delete_webhook.rs
@@ -1,10 +1,10 @@
-use crate::application::http::server::api_entities::api_error::ApiErrorResponse;
-use crate::application::http::server::api_entities::{api_error::ApiError, response::Response};
-use crate::application::http::server::app_state::AppState;
 use axum::{
     Extension,
     extract::{Path, State},
 };
+use ferriskey_api_core::api_entities::api_error::ApiErrorResponse;
+use ferriskey_api_core::api_entities::{api_error::ApiError, response::Response};
+use ferriskey_api_core::app_state::AppState;
 use ferriskey_core::domain::authentication::value_objects::Identity;
 use ferriskey_core::domain::webhook::ports::{DeleteWebhookInput, WebhookService};
 use serde::{Deserialize, Serialize};
