+++ libs/ferriskey-api-webhook/src/handlers/fetch_webhook.rs
@@ -1,10 +1,10 @@
-use crate::application::http::server::api_entities::api_error::{ApiError, ApiErrorResponse};
-use crate::application::http::server::api_entities::response::Response;
-use crate::application::http::server::app_state::AppState;
 use axum::{
     Extension,
     extract::{Path, State},
 };
+use ferriskey_api_core::api_entities::api_error::{ApiError, ApiErrorResponse};
+use ferriskey_api_core::api_entities::response::Response;
+use ferriskey_api_core::app_state::AppState;
 use ferriskey_core::domain::webhook::entities::webhook::Webhook;
 use ferriskey_core::domain::webhook::ports::WebhookService;
 use ferriskey_core::domain::{
