+++ libs/ferriskey-api-webhook/src/router.rs
@@ -3,7 +3,8 @@ use super::handlers::delete_webhook::{__path_delete_webhook, delete_webhook};
 use super::handlers::fetch_webhook::{__path_fetch_webhooks, fetch_webhooks};
 use super::handlers::get_webhook::{__path_get_webhook, get_webhook};
 use super::handlers::update_webhook::{__path_update_webhook, update_webhook};
-use crate::application::{auth::auth, http::server::app_state::AppState};
+use ferriskey_api_core::app_state::AppState;
+use ferriskey_api_core::auth::auth;
 
 use axum::{
     Router, middleware,
